Extract "collidable" properties separately.
Previously these were bundled in with the text properties, which
duplicated several of them across different symbolizers and was an
obstacle to making `point_symbolizer` re-use common placement
code.

In bringing the "collidable" properties together and making them
common across uses of `placement_finder`, it brings support for
"ignore-placement" to TextSymbolizer and "avoid-edges",
"minimum-distance" and "minimum-padding" to PointSymbolizer. The
use of a common implementation should make it more consistent and
easier to modify and maintain.