WIP: refactoring `markers_symbolizer` layout.
Re-using existing layout code in `text_symbolizer_helper` for
laying out `markers_symbolizer`. The layout is generally very
similar, but has been re-implemented and extended in some ways for
`markers_symbolizer`.

The work in progress so far gets some things right, but is missing
some scaling for SVG markers, and some of the "multi" layout
options.